Add display totals to groups page

+++ db/migrate/20230220141947_create_accounts.rb
@@ -2,7 +2,7 @@ class CreateAccounts < ActiveRecord::Migration[7.0]
   def change
     create_table :accounts do |t|
       t.string :name
-      t.integer :amount
+      t.decimal :amount
 
       t.timestamps
     end
